Rewrote linkage parsing.  Now parses linkages from tables in Chinese characters correctly and handles many other linkages much better.
Implemented heuristic classification of parenthesized parts for linkages
(and the same code can later be used for translations).  It tries to guess
whether the parenthesized part is tags, romanization, taxonomic name,
English text, or something else (e.g., kanji spelling of the word).  There
were some changes in the data format generated for linkages.

Fixed some bugs in tag parsing (capitalized words from the beginning of
tags were sometimes ignored).

Various other changes.  This is overall a big chunk of changes and
it is quite possible this breaks something, particularly in linkage
parsing.  More testing is needed in the coming days.