Daily dose of Arabic number paradigm names, and some changes. The number of words with errors had dropped down under 0.1% this morning, so this is just basically clean up by this point and hunting down stragglers. The remaining errors will be from certain specific tables that are hard to parse.
I've decomposed the "-ah" and "-an" etc. forms into tags that are separate
from the *tote and number tags. There started to be too many of them,
and it's apparent that they're not coupled together on that level, but
that they're just different word form paradigms that can be used
for many different morphemes.

The new tags have the form ar-infl-***. The *** is replaced by the
romanized ending "(-in)" or "(-ah") style, except for "(-an)", which
is used for two different forms that I've given the tags ar-infl-an-
maksura and ar-infl-an-alef for the characters used in the Arabic
script.