Implemented debug messages and the set LANGUAGES_WITH_CELLS_AS_HEADERS.
This code doesn't do anything yet beyond print out debug messages. The debug data will
be used to populate LANGUAGES_WITH_CELLS_AS_HEADERS, and when that's far enough it
will be enabled as a kind of check-this-language permission list for a few heuristics.

Some languages on Wiktionary have badly formatted tables, from a machine reading point
of view. Usually what should be table header cells (<th> in HTML output) are normal
cells, often mixed with correct header cells horizontally on top, or missing
altogether (or, rarely, a td-cell is typoed as a th-cell).

This is why there are some heuristical bits of code in inflection.py
handle_wikitext_table() that tries to determine whether a cell should actually be a
header.

One of them checks whether the cell in question receives any kind of tag starting with
"error-" when run through expand_header() the cell in question receives any kind of
tag starting with "error-" when run through expand_header(), which parses header cells
and outputs the tags used for cells in that row or column. If there are no tags
starting with error-, then it is a good indicator the cell is actually a header,
because a normal cell's contents would trigger some kind of tag parsing error.

However, this causes problems when a normal cell contains something found in inflmap,
like personal pronouns (previous bugs were that it could happen even with pronouns in
another language) or straight-up words that are grammatical terminology or have the
same form as gramatical terminology. Also in languages with badly formatted tables,
when a table header (that is cell element) contains unknown header material (not in
inflmap, basicall), then the whole table is borked, but that is less of an issue
because it should be corrected in any case.

The second heuristic is to check whether the table cell has the same style as the
first cell of a row. This specific test might actually be robust enough to be left
alone, but at the moment there is still a debug message.

Other heuristic tests (like the one that tests whether a cell takes up the whole
rowspan of a row, which should reasonably always be a header)
are robust enough to not prompt these debug messages (yet).

At the moment LANGUAGES_WITH_CELLS_AS_HEADERS is not yet populated, except for Greek,
which has a TON of small separate tables that have this kind of incorrect table header
formatting. As time goes by and the set of languages becomes populated, I will enable
a few lines of commented conditionals in the code that will prevent the above
heuristics from happening in languages not in the list. This should remove a bunch of
minor one-off bugs found in many, many languages. Languages in the set will be affected by the
previous behavior and have their cells determined by the tests.

HOW TO FIX ON WIKTIONARY:

If you have a language that could be in LANGUAGES_WITH_CELLS_AS_HEADERS, but also
could have its tables corrected on Wiktionary, please do that before adding the
language into LANGUAGES_WITH_CELLS_AS_HEADERS. This is the simplest to do with
languages that have only a few short tables to correct: find lines in the
Template:lang-something-template code (in Edit on that page), and replace |-pipe
symbols at the start of lines with !-exclamation point symbols, when that specific
line only contains data from the header cell. Preview it, and then check the preview
with your browser's DOM inspection tool (right-click-menu Inspect in Firefox, for
example) and figure out whether the cell you changed is now a <th>-element or not.

Changes made into tables on Wiktionary will, as usually, propagate to kaikki.org and
your own instance of wiktextract only after a new Wiktionary data dump comes out with
the update table.