Added tag transformations based on header direction.
Some tables use whether a header is in a row or column to determine the meaning of
that header, in our case what tag should be output. For example in Swahili contexts
"c5" can mean "class-5" *or* "object-class-5" because it's at the start of a row in
one of the giant tables Swahili has. This is not textually indicated anywhere, only
contextually, but within those tables it is consistent that rows are for
object-concord and columns are for subject concord.

I've added a new subfile with two dicts, rowtag_replacements and coltag_replacements.
At around line inflection.py/2500 there is a part where the tags of a row and the tags
of a column are combined together for the tags of a form. I've added here a simple
check and replacement-function call that will look at whether the language processed
has a dict entry in Xtag_replacements, and then replaces any tag strings with ones
from the replacement map as appropriate.

Use of this functionality can be lead to errors, especially if a language has many
kinds of tables. If the meaning of a "row" header is different in two of these tables
for the same language, one of them will be erroneous in all cases (as it also is when
NOT using this functionality, but still).