Fixed: bold and italics interfering with literal apostrophes.
In some (rare?) cases, text with mixed bold or italics tags together with a literal
apostrophe (') would cause the apostrophe to not appear: `'''C''''est` would render as
`Cest`.

The cause was eventually that the text was used in a template:`{{Q|...|'''C''''est}}`,
which does not parse bold or italics wikitext.

Then this template argument was fed into a module, which generated *raw HTML code*:
`<span>'''C''''est</span>`

This wikitext formatting was automatically cleaned up by a non-parsing simple regex in
wiktextract/clean.py, which has now been replaced by replicating the functionality of
token_iter() in wikitextprocessor, which correctly parses bold and italics tokens.

This took me too long.