Removed duplicate key entries in inflectiondata.py/inflmap.
inflmap is hand-written, and python does not have, and will not have, any native
warning system for duplicate keys overwriting each other in dicts.

By going through the raw python code strings with a script, I manually removed found
duplicates, which lead to a bug with Lithuanian tables in `make test`.

After searching and hunting for a bug in parse_simple_tables that would have been
caused by me breaking up a large for-loop into pseudo-subfunctions, it eventually
turned out that the issue was caused simply by planned behavior: the header 'jūs' that
was being expressed as having a "singular"-tag instead of "plural" was doing so
because of its inflmap rule, and it did so correctly. The inflmap rule was using
base_tags to determine what kind of tags would trigger the "if"-key rule in inflmap
for jūs, and it *did* contain "singular". This "singular" was inherited from the three
first header cells in the row, and fixing the issue meant just removing "singular"
from them, because they were not 'meaningful' headers, only repeating the same data
that was already present.