Detect table nesting depth with infl_map rules
Added new condition "nested-table-depth" that detects how deep, starting
with handle_generic_table, we are when parsing a table.

nested-table-depth can be an int or a list, and refers to the depth
level (0 = root of the table tree), which helps distinguish certain
headers used in different contexts, like class headers in Swahili
megatables.

So if "nested-table-depth" condition checks for "0", if the header is
found in at the top level and not a subtable, it checks true.