Finalized save-and-restore hdrspan system for subtables
Some languages have sprawling megatables (Swahili) where the headers
from the top of the outer-most table (the "root" table) do not
propagate down into lower rows because subtables were and separate
tables that did not communicate with others.

Using the new dummy-tags `dummy-store-hdrspan`, `dummy-load-stored-
hdsrpans` and `dummy-reset-stored-hdrspans` tables within the same table
(in the same outer handle_generic_table call) can now propagate
downwards by explicitly saving a headerspan by having
dummy-store-hdrspan in tags. Then the stored hdrspans can be loaded into
`hdrspans` as 'normal' hdsrspans at another infl_map rule with
dummy-load-stored-hdrspans, and emptied with
dummy-reset-stored-hdsrpans.

"Downwards" in this case means 1) outer towards inner subtables and 2)
tables parsed before into tables parsed later.

This means that even if an outer table actually has content that is
separated by subtables that are in-between, it can only propagate into
those subtables and content that is "later" cannot be propagated into from
subtables. This is an improvement on how it was before, because in earlier
versions of this code the outer tables were actually parsed after all
subtables because of recursion.