Split parsing of different templates
In parse_inflection, when a section that is to be parsed contains
several templates ("{{template}}"), that text is split into separate
sections that are parsed in separate passes.

parse("first {{template1}} after1 {{template2}} after2")
 =>
parse("first {{template1}} after1")
parse("{{template2}} after2")

Doing it this way does not seem to actually change the outcome of the
result of parsing, because the data received from the parser is just
passed on into `data` and mixed together. Diffing a before and after
case showed identical json files as results.

This is because top-level templates are independent of each other and do
not depend on other templates. They are pure functions.

As for text that is not part of a template signature, those are appended
in a string together with the template immediately preceding them. The
only exception is text before the first template, which 'belongs' to the
first template.

The parser uses `after` variables for text after a table, which is
why they are joined in this order.