Finalized flattened_tree()
The previous version of flattened_tree() was pretty simple and only
handled lists after another, and not hierarchical lists.

This new version recursively iterates through the section content,
and yields each node as it encounters them.

Sublists, lists that are deeper nested, are in our code coded as *a*
member of a LIST_ITEM's children, but it's not necessarily the last
entity (as it should probably be) because of hanging strings ending up
there.

Currently these sublists are handled by roughly separating them from
their parent node and doing them as a separate recursion immediately
after the main node (which has the sublist removed from its children).

This *shouldn't* break because of >1 sublists; there should only be one
possible sublist for a LIST_ITEM.