Filtering of Pronunciation sub-sections by POS
When parse_pronunciations() encounters a line with one or two words that
correspond to a key in part_of_speech_map, it toggles active_pos to be
that POS and assigns it to "pos"-fields in values of "sound".

This value is later used in page/parse_language()/merge_base() to filter
out values with a "pos" field that doesn't match the POS of the entry
the data is merged into. The POS-field is then removed from the value,
its job done.

This can be expanded: the POS-determining function is currently the dumbest
possible way I could think of, but could be elaborated with a better regex
and maybe using decode_tags (although that's a bit costly), and in the same
place you could also added other kinds of filters, as long as you mirror
them in merge_base().