Fixed #168, added new "head_nr" field to glosses and forms
Previously, when a part of speech section had several head sections one
after another, later sections were not parsed.

For the part of glosses, this was a bug where a line had "break" instead
of "continue", meaning that when a list was encountered, it was added to
a list of lists and then the loop broke, meaning a list would only even
contain one list of glosses.

For heads, it was never considered that there might be several templates
in the same PoS section.

These are now parsed, and each "section" of a head (and maybe list of
glosses) is handled such that they have a 1-indexed number that is added
to forms and glosses so that the data has at least some indication of
where a form or gloss belongs to.