Fixed possible issue with splitting inflection templates
In an earlier version of this code, we tried to split up blocks
containing top-level templates (like separate tables) by counting pairs
of curly braces ("{{ }}") and their nesting to determine whether we've
exited a top-level template.

There is the theoretical possibility of triple braces ("{{{}}}") that
could break the nesting counter, which is fixed here by counting single
braces, as long as they're in a contiguous string of two or more braces.

This should work because regardless of how double and triple braces are
nested, the number of curly brace *characters* should also be balanced
through the top-level template. We do not have to care about the
structure those braces created within the template, because we're just
sending everything onwards into another function.

Example:

"{{test {{{1}}}}}" => ["{{", "test ", "{{{", "1", "}}}}}"]

"{{" -> nesting +2 = 2
"test"
"{{{" -> nesting +3 = 5
"1"
"}}}}}" -> nesting -5 = 0
End of template!