Using --language now skips pages without target languages
Previously when wiktwords was used to parse languages that were not
ALL_LANGUAGES every single page in the cache or dump file was parsed,
and only then was the parse tree checked if it contained a
language-specific section.

Now, as a preprocessing step the text of the page is searched and
if "== Language Name|Other Language|Alternative Name" is not found
the page is ignored and not parsed.

This should drop the processing time of languages by a couple of orders
of magnitude, more for smaller languages and less for big languages that
have words in many articles.