Filtered out "Category:" when parsing heads
Previously if a part-of-speech section (head + sense glosses) ends an
article and there are Category: link templates at the end of it,
the text of the "Category:..." became part of the head section being
parsed, even after a list of glosses, because the section was never
terminated by another section.

Currently the in parse_part_of_speech works roughly like this:
  - loop through the children of the POS node
  - append found LIST nodes to a list of LISTS (glosses)
  - append everything else into pre, which should contain heads
  - if you encounter some kind TEMPLATE (assume it's a head template):
    - start a new subsection of POS (recently implemented)

Then these subsections are handled (and assigned the new head_nr field
if applicable). `pre` contains lists of nodes and text, the 'head'
portion of the subsection, but the stuff *before* the glosses and after
the glosses is not separated... Which might be the correct way to do
this.

The kludge is that when going through the lists within pre, if a string
starting with "Category:" is encountered, discard it.

XXX: Implement a better solution.