Stop collecting word head data at a list
Previously, word head data was collected as a string in a part of speech
even when logically it should not have been after a list of senses.

Now, when encountering a list of any kind, the current 'head' is not
added into anymore until the next 'head' segment starts at a template,
'\n\n' or '<br>'.