Fixed some overzealous head-template separation
With a previous commit to fix minor issues with PoS sections that had
several head templates (that it couldn't handle) a logic bug was
introduced that caused, as was intended, new "head" sections to start at
top-level templates.

However, it is not at all rare that a head template is followed by a
qualifying template like {{g}}! This broke up the line with the not-head
template into a "head" section, creating a new erroneous form or other
garbage data.

This fix adds checking whether the template that is being inspected is
the first template of a paragraph. If it is, we assume it is a head
template (could be something like a table, of course), and handle it as
previously.

tl;dr templates create new head sections only once per paragraph.

XXX for tomorrow: it shouldn't be "once per paragraph", should be
"immediately at the start of a paragraph"; text followed by a template
will break this.