infl_map rules and a quick pylint script
infl_map has a problem as a human-written dict: duplicate keys are not
uncommon.

Python dicts do not have any kind of warning message possibility within
python that would tell when a key is overwritten (in the same piece of
assignment code, even!), and JSON objects are not compatible with the
trailing spaces we use in our python dicts and the comments. Taking the
current dict and scanning it for keys manually is not possible without
*actually* parsing the python code, and it can't be converted to JSON
and back (the json module can be used to check for duplicate keys in
json data), and regexing is extremely ill-advised with nested structures
like dicts.

Instead of doing runtime checking, the next simplest is to use pylint,
which will inform us about duplicate keys with all messages except for
W0109 messages disabled.

inflmapduplicatekeys.sh is just
$ pylint  --disable=all --enable=W0109 wiktextract/inflectiondata.py

and is saved here because it will disappear from my head the moment
I press ^S ^X and commit this.