Fixes #182; xlat_head_map tag overlaps with article title
Some keys in xlat_head_map, which contains data about text describing
what the head of a word is, like "archaic", "m" or "nominative", can
overlap the title/form of a word like happened in the case of #182. When
this happened, forms like "more defunct" was interpreted as "more", with
"defunct" only adding the "archaic" tag because it was used that way in
Norwegian entries.

There was previously a test to check for this kind of overlap in the
form of `not ctx.title.endswith(" " + tagkeys)` (where `tagkeys` was the
xlat_head_map key in question), but it failed when the whole title (like
"defunct") was equal to the key.

I've added a complementary test that should take care of this edge case.

HOWEVER XXX: this does not handle well cases where the keys from
xlat_head_map are NOT equal to ctx.title (article main head word form),
BUT the word has a form appearing in the head that is equal to something
in xlat_head_map. Hypothetical example: "defunc, mf, accusative defunct"
could break.