Added "Internet slang" in uppercase_tags
uppercase_tags contains words and phrases that are usually to be
parsed as locations or dialects, or, in this case, slang. Previously,
"Internet slang" was parsed as the tags ["Internet", "slang"], now
it has the most specific tag "Internet-slang" as generated from
uppercase_tags.

Also removed two debug prints I forgot to remove.