Fixes #186, titles with commas with wrong canonicals
In phrases with commas, the commas were used in regexes to split up
the phrase into two different canonical forms. There is now a kludge
to fix this by splitting and ignoring the phrase with one or more
commas, if ctx.title contains a comma.

The other fix down is because the comma is eaten by a regex used to
process text in the head, creating a canonical form without the comma as
the split words are reconstituted in add_related.

As a reminder, the expected behavior is that if the head has the exact
same form as the title in it as a form, it is not expressed separately.
Only when there's a different kind of form (possibly spelling variants)
does the json data contain a separate "canonical"-tagged entry. I think
this happens in add_related or later, so should not be filtered in the
part of the code that has been modified here.