Added some template filtering in head parsing
Some templates, like {{category}} should typically only appear at the
end of a section, and are not part of head, so parsing a head section
should end when encountering it.

There is already a panel_template list used to skip certain templates
that generate floating panels that aren't part of the text they're in,
so we should also ignore those (but not stop parsing the head section).