If not LIST in Pronunciation, expand templates
Fixes #192

Issue: if a Pronunciation section does not have a top-level WikiNode
List, but instead it is generated inside a template, like in Italian
or Finnish entries, it did not parse at all.

Solution: if the contents being looked at does not contain any
LISTs, expand and extend all the templates into a new list.

Also cleaned up some of the output text of "* " at the start.