Handled {{see translation subpage}} without parameters
Before, if {{see translation subpage}} was without arguments (that is,
no POS data usually), the default was to return and put out an error.

This behavior has been changed so that if there is no argument in the
template, we check if there was and send out a path sequence using
ctx.subsection (= current POS) by default.