{{see translation subpage}} now defaults to language
The previous commit made argumentless {{see...}} templates
default to using the current PoS instead (ctx.subsection), which
is technically not correct and different from Wiktionary behavior.
Because Wiktionary defaults to linking to a Language section,
we will follow that and trust in Wiktionary users keeping
{{see...}} templates correct when there are several present
on a translation subpage. Currently, as long as there aren't
any other PoS sections on the translation subpage, there is no
cross-contamination.