{{see translation page}} links with "Etymology X: PoS"
Some {{see translation page}} templates have to link to subpages with
numbered Etymology sections, not just language and pos as usual. Afaict
there is no normal syntax even on wiktionary to do this, so for example
port [English] has resorted to stuff like "Etymology 1: Noun", which
creates a broken anchor, but at least links to the right *page*.

Fixes #200, by detecting "Etymology X: PoS" with regex and putting
that in seq.