Fixed: gloss has examples and subglosses
Previously, when a gloss had subglosses that were succesfully parsed
and added to pos_data, the higher up gloss (supergloss) would be
skipped as redundant, because its data had been propagated downstream
already. However, examples were not part of this, and if a supergloss
had examples and was skipped so too would those examples.

Now, when a supergloss has examples it is always parsed, even when
a subgloss returns as parsed.