Fixed unextracted categories in examples
Originally, before the change in this branch to purely recursive
sense list parsing, the behavior of extracting categories like
"English terms with usage examples" was erratic and seemingly
random.

This was caused by the template function used in parse_sense()
skipping over a bunch of difference {{ux}}-related templates completely,
and handling examples in a separate loop.

However, if a ux-template was inside some other entity, like LIST_ITEM,
then it would get parsed as usual, and the category meta-data would
slip through.

After switching the recursive version in this branch, even the few
that randomly slipped through previously did not, so it seemed like
the new version had a bug where it just couldn't find the category
data for certain things, because it did the separation of examples
in a more strict way than earlier.

In other words, it wasn't my fault, there was a bug that looked like
a feature!

The fix here is to collect meta-data into sense_base also inside
the example extraction function.