Kludge to remove German articles
Sometimes, a previous kludge is just not enough and you
need to take a shotgun to some German articles floating
in a table.

Example: eulersche Formel/German in Jan 2023 at least

If a language has "conditionally_ignored_cells" in its lang_specific
configs (and German's article-removing kludge doesn't activate) then:

If a table has words that can be found in the list values
of c_i_c, skip over that cell, but only if:
    1. the key of that list is not in the tags of the
       the cell (for example "definite" for "die")
    2. the word article is found in that same list

The second condition is for tables in articles for those words
themselves, and related words in the list, like articles.