Added wiktextract/usertools folder
Starting with two scripts for sorting json output files and
a third to search for words in a json file, this folder
is meant to hold scripts for end-users who want to manipulate
and browse through the data.

The sorting scripts are slow because the files they sort
are slow, but they should generate files that have
a predictable sorting key based on the json data itself,
and thus create files that can be compared to each other
with tools like diff.

The word-search has a few more parameters, including filtering
with regex, by language(s) and max count for output. This
just throws hits into stdout.