Added script to compare random samples; bugfixes
Added to usertools json_compare_samples.py. This is a tool that takes
two versions of the same Wiktextract json data, takes a random sample
between them and compares those randomly chosen samples using difflib.
Output is basically a bunch of diffs between the compared objects (using
basic string diff because the other way seems like a weeks-long hassle
that ends in tears), so you can quickly see changes.

The script takes as its parameters two files, with the "same" data, and
optionally --once-in-a N (default 10000) that is just a 1/N frequency
for taking samples line by line. Each line, roll a die with ten thousand
sides, if it rolls 0 then pick that line as the random sample; if you
put this parameter as 1 it will pick every line.

Fixed a bug where the "english" field of senses was duplicated,
introduced by the new recursive code when parsing sense lists.

The "english" field in a sense is a misnomer and will be changed
to the more appropriate "qualifier".