parse_sense_qualifier() now adds "qualifier" field
Previously, parse_sense_qualifier() would add an "english" field when
parsing sense qualifiers, mainly due to inertia with how classify_desc()
is used (which returns a value "english" when the text it classifies is
English text and not tags or topics), but the field-name "english"
doesn't make sense in the context of a sense object.

Depending on the need, "qualifier" could be expanded into other places
in the data scheme, but as it is, at a quick glance most other examples
of "english" are appropriate, translations or glosses.