Fixed bug in decode_tags
Not exactly sure what I fixed here, but I found a small bug with a for loop
that, when corrected, fixed an issue we had.

Bug: if you had "bar f or foo m", while xlat_tags_map had any string starting
with "m [any character(s)...]", parsing "foo m" (because of the "m" would fail.

This was because decode_tags failed to return tag data with
["error-unknown-tag"] in it. The culprit to this was found in a for loop that
iterated over "path" taken from pos_paths[start_i], but there was no test to
check if pos_paths[start_i] was in fact an *empty* list. Inside this for-loop
the tag data with the error tag data and a duplicate of `path`'s data was
supposed to be appended (into pos_paths[-1]), but this didn't happen when
pos_paths[start_i] was empty.

I corrected this by checking if pos_paths[start_i] is empty, and if not,
appending `u` without appending `path` (which doesn't exist).