Fixed issue with zh-pron templates not parsing
In a previous commit, I had "unrolled" Pronunciation sections that
did not contains LISTs so that all to-level templates were expanded
(and thus actually parsed, if they contained the LIST that was
missing).

However, this broke the special handling of the zh-pron template used
in Chinese entries (mostly character pronunciation stuff), which
contains reflex pronunciations of the character(s) in different
Chinese languages and their dialects.

This has been corrected by checking whether the template being
expanded isn't actually "zh-pron".