Three-line original/romanization/translation examples
Added an if branch to take care of examples that have the structure

ORIGINAL TEXT
Original text romanization
Translated into English

when the number of lines is exactly three, which was not a case
considered before this. This is to handle ja-usex example templates
with this strucutre, if other similar forms turn up they should
be added to the control flow in extract_examples() alongside
this.