Added "ruby" fields to forms and *nyms
Added some new code that adds a `ruby` field to forms and some linkage
items (where it was already handled partly, forms did not have ruby
stuff).

"Ruby" is the term for when Japanese text (almost always kanji) has some
smaller hiragana or katakana (syllabic/phonetic) script on top of it to
indicate the pronunciation of the text. This is done in html with <ruby>
elements, inside which is the main text, <rp> elements containing
discarded (or displayed) parentheses (like img alt text) and an <rt>
element for the prounciation guide text (furigana).

To get the ruby data as perfect as possible, I've actually parsed the
HTML and then extracted tuple pairs of (text[kanji], ruby[furigana]),
that is appended to ruby-data.

In parse_word_head, we pass on ruby-data, and when parse_word_head
encounters text that fits that data ("kanji(ruby)") it adds two
temporary magical parenthese __lrub__ and __rrub__ into the text
so that there are no collissions with other paren-seeking regex
patterns.

Then when we handle word forms in add_related, we look for those
__lrub__ __rrub__ parentheses, add the ruby data (extracted in-situ,
it's been scrambled by this point because we should be able to trust the
pairings because they've already been handled in parse_word_head) into
the form or alt-of data, and also remove the ruby from the main-text.

The handling of ruby in the *nym templates is a bit simpler, because
we have actual HTML-nodes still left to handle within the *nym "item",
so we don't have to do any of the silly magical parentheses stuff;
the template being handled results in one data-item, so we just
extract the ruby data, remove the ruby text and give the data to the
main item (synonym, antonym etc.)