Fixed a bug in classify_desc()
Previously when doing unicode data stuff with the text,
the unicode data was decomposed into components pieces
(like an ellipse "…" -> "...") to make it easier to figure
out what data they have (especially with COMBINING characters),
but that data was then zip()ed together with the original text,
causing mismatches when handling it; the indexes were off, so
"character -> unicodedata" did not hold true.

The description data is now normalized into normalized_desc,
which is used when appropriate to analyze characters.