Handle examples with the format ref, original, romanization, english
Some Japanese examples have a quotation as their examples in the format
quotation reference, original Japanese text, romanization and finally
English translation, but there was no path in extract_examples to
extract the romanization portion of the text, which was left in
`lines` and became the tail part of `text` in the example.

Added a check to see if a line looks like romanization for classify_desc(),
after the English translation and references have been extract, then
add it in `roman`.