XXX use the new parameter in ctx.process()
We want to intercept the template az-suffix-form because it creates
a floating div that either gets removed completely or messes up
if it's not removed. We want to intercept it, but currently it is
pre-expanded when wiktionary is parsed in wikitextprocessor, so the
template has just become a nameless table that we are unable to identify.

In wikitextprocessor, I've added new parameters to Wtp.process() that
ultimately downstream acts as a whitelist and blacklist for Wtp.need_pre_expand
stuff (XXX), so we can forbid the expansion of the az-suffix-templs
template and intercept it by extracting it, and then using that template
some other way (as part of the Conjugation section etc.)