Simplified path of do_not_pre_expand stuff
Turns out that even though Wtp.need_pre_expand (ctx.need_pre_expand)
is generated in wikitextprocessor and affects the cache, it doesn't
do so by actually pre-expanding anything into the wikt-cache file
but only by creating the set of templates that would later
be used when parsing with Wtp.parse and Wtp.expand!

This means that instead of always having to recreate the cache,
or having to pre-fill Wtp.need_pre_expand ahead of parsing, we can
just inject or remove stuff later from that set before parsing
(with pre_expand=True) happens.

... And this was already being done! That's what additional_expand
was for. And following the breadcrumbs of that parameter, we can
modify Wtp.expand in a similar way we did before with Wtp.need_pre_expand
(but on the fly, so to speak), bypassing a detour through wiktionary.py
etc.