Fixed issue with returned functions when extracting categories
See previous commit; in that commit, we only used tostring() to get unbreak
the code when receiving a function where we expected a string.

Turns out, the descriptions we assume are strings are more like templater
object: a string with stuff like {{{langname}}} or in this case an actual
function that would be later called with something like a context data
table that would supply it with its parameters (like language). The
`gerunds` function is the only one that exists right now and it handles an
empty table gracefully (failing first and returning the default template),
so we just call the function with an empty table and it kind of works

Left a few prints in place, the category extraction stuff happens only when
a categories_file parameter is supplied to wiktwords, so typically only
when running the whole thing.

HOWEVER A TIP: if you want to test extract_categories code, you *can* just
call wiktwords on a single page (any page) and if you have
--categories-file in the parameters it will run extract_categories. You
need a cache for that, but you would have it anyway. V. handy for testing.