New "dummy-section-header" trigger in inflmap
The new Swahili not-so-mega tables have a problem in that there are too
many <table> elements that break the flow of our parser, a bit needlessly.
When "Infinitives" in a separate <table>...Infinitives</table> from the
data it should be heading (<table>...positive: blah...</table> it can't
propagate, even with the old header-saving subsystem devised for the
original Swahili megatable (see those commits).

To fix this, using TableContext we introduce a new context-variable called
TableContext.section_header which should contain the tag data of the
current "section" being parsed, within a table with subtables.

When inflmap finds the dummy-tag "dummy-section-header" it will save the
tag data (of only one, the first) where it encounters it in a sect of tag
alternatives into TableContext.section_header, and then later joins that
data in the data of the cells.

When it encounters dummy-section-header the next time, it will replace the
tag data with the new section data.