Handle broken ruby elements better
Sometimes a template creates a ghost ruby-element (which is removed
later by wiktionary or which we accidentally generate in our own process)
where some part of the ruby main or ruby element is missing: in these
cases, parse_ruby should return None, and when calling parse_ruby
it should be checked whether something is actually returned.

This fixes two articles where we had __lrub__ __rrub__ magical strings,
out of a whole three in the whole extract.json data.

The last one is some kind of parsing error in a different part of the code,
tables I think.