Fixed broken "gloss promotion" kludge and changed raw_glosses deletion
One bug:

When the code encounters a gloss with a single subgloss

`# main gloss
`## subgloss

we have kludged this so that the the subgloss is "promoted" out
of the hierarchy to the same level as the main gloss, because
this seems like a mistake on wiktionary.

When creating the recursive persion of sense parsing, this behavior
was left in place, but it was a bit broken because if the main gloss
was something like "(intransitive)", it would result in an "empty"
gloss that was not pushed into sense_data, and thus sense_data would
not be reset. This caused duplicated data if there was something
in sense_base that was being fed into the function call for
the subgloss. This has been fixed by manually emptying sense_data.

One change:

Previously I changed raw_glosses data to be deleted if it could be
found in `glosses`, which was to my mind unnecessary duplication.
However, because raw_glosses and glosses are both lists, you can
have them actually be out-of-synch which makes it harder to use
glosses-data instead of raw_glosses data when needed (what goes where??),
so a change has been made to just compare if raw_glosses and glosses
are completely identical and have the same length, and then delete
raw_glosses if this is so. This way, you can always rely that
glosses can be a drop-in replacement for raw_glosses when raw_glosses
is not found.